+++ src/lib/components/copy/CopyButton.tsx
@@ -13,6 +13,8 @@ interface CopyButtonProps extends ButtonProps {
   hasIcon?: boolean;
   buttonText?: string;
   amptrackSection?: string;
+  amptrackSubSection?: string;
+  amptrackInfo?: string;
   iconGap?: number;
 }
 
@@ -25,6 +27,8 @@ export const CopyButton = ({
   variant = "outline-accent",
   buttonText = "Copy",
   amptrackSection,
+  amptrackSubSection,
+  amptrackInfo,
   ml,
   iconGap,
   ...buttonProps
@@ -41,7 +45,11 @@ export const CopyButton = ({
         size={size}
         float="right"
         onClick={() =>
-          track(AmpEvent.USE_COPY_BUTTON, { section: amptrackSection })
+          track(AmpEvent.USE_COPY_BUTTON, {
+            section: amptrackSection,
+            subSection: amptrackSubSection,
+            info: amptrackInfo,
+          })
         }
         {...buttonProps}
         borderRadius={size === "xs" ? 6 : 8}