+++ src/lib/components/table/TableTitle.tsx
@@ -1,17 +1,17 @@
 import type { BoxProps } from "@chakra-ui/react";
 import { Badge, Box, Flex, Heading, Text } from "@chakra-ui/react";
 
-import type { Option } from "lib/types";
-
 interface TableTitleProps extends BoxProps {
   title: string;
-  count: Option<number>;
+  count?: number;
+  showCount?: boolean;
   helperText?: string;
 }
 
 export const TableTitle = ({
   title,
   count,
+  showCount = true,
   helperText,
   mb = 6,
   ...props
@@ -21,9 +21,11 @@ export const TableTitle = ({
       <Heading as="h6" variant="h6">
         {title}
       </Heading>
-      <Badge textColor={count ? "text.main" : "text.disabled"}>
-        {count ?? "N/A"}
-      </Badge>
+      {showCount && (
+        <Badge textColor={count ? "text.main" : "text.disabled"}>
+          {count ?? "N/A"}
+        </Badge>
+      )}
     </Flex>
     <Text variant="body2" textColor="text.dark" fontWeight={600}>
       {helperText}