+++ src/lib/types/tx/transaction.ts
@@ -50,7 +50,7 @@ export interface BaseTxFilters {
   isIbc: boolean;
 }
 export interface WasmTxFilters {
-  isUpload: boolean;
+  isStoreCode: boolean;
   isInstantiate: boolean;
   isExecute: boolean;
   isMigrate: boolean;